Audio_Intent/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.21.0
librosa>=0.9.0
soundfile>=0.10.0
sounddevice>=0.4.0
tensorflow>=2.10.0
onnxruntime>=1.12.0
pyopenssl>=23.0.0
